Use only Satellites with Transmitters on fixtures
* also move tests methods
